The response named RoomTypeResponse has been changed to RoomTypesResponse and the response named RoomTypeBasicResponse has been changed to RoomTypeResponse.
