+++ RoomService/src/main/java/com/flz/model/mapper/RoomTypeEntityToBasicResponseMapper.java
@@ -1,7 +1,7 @@
 package com.flz.model.mapper;
 
 import com.flz.model.entity.RoomTypeEntity;
-import com.flz.model.response.RoomTypeBasicResponse;
+import com.flz.model.response.RoomTypeResponse;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,8 +11,8 @@ public final class RoomTypeEntityToBasicResponseMapper {
     private RoomTypeEntityToBasicResponseMapper() {
     }
 
-    public static RoomTypeBasicResponse map(RoomTypeEntity roomTypeEntity) {
-        return RoomTypeBasicResponse.builder()
+    public static RoomTypeResponse map(RoomTypeEntity roomTypeEntity) {
+        return RoomTypeResponse.builder()
                 .id(roomTypeEntity.getId())
                 .name(roomTypeEntity.getName())
                 .price(roomTypeEntity.getPrice())
@@ -21,12 +21,12 @@ public final class RoomTypeEntityToBasicResponseMapper {
                 .build();
     }
 
-    public static List<RoomTypeBasicResponse> map(List<RoomTypeEntity> roomTypeEntities) {
-        List<RoomTypeBasicResponse> roomTypeBasicResponses = new ArrayList<>();
+    public static List<RoomTypeResponse> map(List<RoomTypeEntity> roomTypeEntities) {
+        List<RoomTypeResponse> roomTypeRespons = new ArrayList<>();
         for (RoomTypeEntity roomTypeEntity : roomTypeEntities) {
-            RoomTypeBasicResponse roomTypeBasicResponse = map(roomTypeEntity);
-            roomTypeBasicResponses.add(roomTypeBasicResponse);
+            RoomTypeResponse roomTypeResponse = map(roomTypeEntity);
+            roomTypeRespons.add(roomTypeResponse);
         }
-        return roomTypeBasicResponses;
+        return roomTypeRespons;
     }
 }
