+++ RoomService/src/main/java/com/flz/model/mapper/RoomTypeEntityToResponseMapper.java
@@ -2,7 +2,7 @@ package com.flz.model.mapper;
 
 import com.flz.model.entity.AssetEntity;
 import com.flz.model.entity.RoomTypeEntity;
-import com.flz.model.response.RoomTypeResponse;
+import com.flz.model.response.RoomTypesResponse;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,19 +12,19 @@ public final class RoomTypeEntityToResponseMapper {
     private RoomTypeEntityToResponseMapper() {
     }
 
-    public static RoomTypeResponse map(RoomTypeEntity roomTypeEntity) {
+    public static RoomTypesResponse map(RoomTypeEntity roomTypeEntity) {
 
-        List<RoomTypeResponse.Asset> assetList = new ArrayList<>();
+        List<RoomTypesResponse.Asset> assetList = new ArrayList<>();
 
         for (AssetEntity assetEntity : roomTypeEntity.getAssets()) {
-            RoomTypeResponse.Asset asset = RoomTypeResponse.Asset.builder()
+            RoomTypesResponse.Asset asset = RoomTypesResponse.Asset.builder()
                     .id(assetEntity.getId())
                     .name(assetEntity.getName())
                     .build();
             assetList.add(asset);
         }
 
-        return RoomTypeResponse.builder()
+        return RoomTypesResponse.builder()
                 .id(roomTypeEntity.getId())
                 .name(roomTypeEntity.getName())
                 .price(roomTypeEntity.getPrice())
