--- RoomService/src/main/java/com/flz/model/response/RoomTypeBasicResponse.java
@@ -1,17 +1,0 @@
-package com.flz.model.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.math.BigDecimal;
-
-@Getter
-@Builder
-public class RoomTypeBasicResponse {
-
-    private Long id;
-    private String name;
-    private BigDecimal price;
-    private Integer personCount;
-    private Integer size;
-}
