+++ RoomService/src/main/java/com/flz/model/response/RoomTypeResponse.java
@@ -2,14 +2,11 @@ package com.flz.model.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.util.List;
 
-@Builder
 @Getter
-@Setter
+@Builder
 public class RoomTypeResponse {
 
     private Long id;
@@ -17,15 +14,4 @@ public class RoomTypeResponse {
     private BigDecimal price;
     private Integer personCount;
     private Integer size;
-    private String description;
-    private List<Asset> assets;
-
-    @Getter
-    @Builder
-    public static class Asset {
-        private Long id;
-        private String name;
-
-    }
-
 }
