+++ RoomService/src/main/java/com/flz/service/RoomTypeService.java
@@ -2,19 +2,19 @@ package com.flz.service;
 
 import com.flz.model.request.RoomTypeCreateRequest;
 import com.flz.model.request.RoomTypeUpdateRequest;
-import com.flz.model.response.RoomTypeBasicResponse;
 import com.flz.model.response.RoomTypeResponse;
+import com.flz.model.response.RoomTypesResponse;
 import com.flz.model.response.RoomTypesSummaryResponse;
 
 import java.util.List;
 
 public interface RoomTypeService {
 
-    List<RoomTypeBasicResponse> findAll();
+    List<RoomTypeResponse> findAll();
 
     List<RoomTypesSummaryResponse> findSummaryAll();
 
-    RoomTypeResponse findById(Long id);
+    RoomTypesResponse findById(Long id);
 
     void create(RoomTypeCreateRequest roomTypeCreateRequest);
